refactor: 메인로직의 인풋 메서드를 인풋 객체에 이동
